feat: 마이 프로필 조회 API response 및 로직 수정

+++ src/main/java/com/follow_me/running_mate/domain/member/entity/Member.java
@@ -76,11 +76,11 @@ public class Member extends BaseEntity {
 
     @Column(nullable = false)
     @Builder.Default
-    private Integer followerCount = 0;
+    private Long followerCount = 0L;
 
     @Column(nullable = false)
     @Builder.Default
-    private Integer followingCount = 0;
+    private Long followingCount = 0L;
 
     @Column(nullable = false)
     @Builder.Default
